+++ ci/build-install-tools.sh
@@ -6,6 +6,7 @@ SOURCE_DIR="`pwd`"
 echo $SOURCE_DIR
 TOOLS_DIR=${SOURCE_DIR}/Tools
 PACKAGE_DIR=${SOURCE_DIR}/Package
+ThirdLibs_DIR=${TOOLS_DIR}/ThirdLibs
 
 if [ ! -d "${TOOLS_DIR}" ]; then
     mkdir ${TOOLS_DIR}
@@ -38,6 +39,17 @@ function function_common()
     #    fi
     #fi
     
+    # Download third libraries
+    if [ -n "$DOWNLOAD_THIRDLIBS_URL" ]; then
+        if [ ! -d ${ThirdLibs_DIR} ]; then
+            mkdir -p ${ThirdLibs_DIR}
+        fi
+        cd ${ThirdLibs_DIR}
+        ThirdLibs_File=third_libs.tar.gz
+        wget -c -nv --no-check-certificate $DOWNLOAD_THIRDLIBS_URL -O $ThirdLibs_File
+        tar xzvf $ThirdLibs_File
+    fi
+
     # Qt qt安装参见：https://github.com/benlau/qtci  
     cd ${TOOLS_DIR}
     if [ "$DOWNLOAD_QT" = "TRUE" ]; then
@@ -59,6 +71,8 @@ function function_common()
             fi
         fi
     fi
+    
+    cd ${SOURCE_DIR}
 }
 
 function install_android()
@@ -144,13 +158,13 @@ function function_android()
     #(sleep 5 ; while true ; do sleep 1 ; printf '\r\n' ; done ) | sudo apt install oracle-java11-installer -qq -y
     
     #sudo apt install oracle-java11-set-default -qq -y
-
-    install_android
-    
     #sudo apt-get install ant -qq -y
     sudo apt-get install libicu-dev -qq -y
-    
+    sudo apt-get install -qq -y libxkbcommon-x11-dev libglu1-mesa-dev
+
     function_common
+    install_android
+
     cd ${SOURCE_DIR}
 }