+++ ChangeLog.md
@@ -3,6 +3,7 @@
   + Modify ci for use ubuntu system qt 
   + Modify install.nsi in windows to add " /norestart"
   + Modify 2020 chinese holidays
+  + Android: add openssl to apk
 
 - v0.3.4
   + Modify chinese holidays directory to user directory